// src/authentication/dto/create-student.dto.ts
export const UserUUID = 'user_documentor_';
export const StudentUUID = 'student_info_';
export class StudentDto {
  username: string;
  password: string;
  email: string;
  name: string;
  title: string;
  signatureurl: string;
  studentcode: string;
  semeter: string;
  academicyear: string;
  faculty: string;
  department: string;
  level: string;
  educationlevel: string;
  course: string;
  stauts: string;
  gpax: number;
  phone: string;
}